requirements.txt
# Core dependencies
agno>=0.1.0
openai>=1.0.0

# Google API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Additional utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0